# src/main/resources/static/swagger/mng-api-docs.yaml
openapi: 3.0.3
info:
  title: 관리자 REST API 명세 ( yaml )
  description: 관리자 REST API 명세 with Swagger ( yaml )
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - Authorization: []
paths:
...
  /api/mng/code/proc:
    put:
      tags:
        - 코드 관리 컨트롤러
      summary: 코드 수정
      operationId: codeUpdate
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/CodeMngVO'
            examples:
              test:
                value:
                  code_seq: '13'
                  code: 'test'
                  code_nm: '테스트 코드'
                  use_yn: 'Y'
                  ord: '1'
                  remark: "테스트 코드"
              guide:
                value:
                  code_seq: '코드 순번'
                  code: '코드'
                  code_nm: '코드 이름'
                  use_yn: '사용여부 (Y ,N)'
                  ord: '정렬 순서'
                  remark: "메모"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
    post:
      tags:
        - 코드 관리 컨트롤러
      summary: 코드 등록
      operationId: codeInsert
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/CodeMngVO'
            examples:
              test:
                value:
                  code : 'test'
                  code_nm : '테스트 코드'
                  prt_seq : '0'
                  use_yn : 'Y'
                  ord : '1'
                  remark : "테스트 코드"
              guide:
                value:
                  code: '코드'
                  code_nm: '코드 이름'
                  prt_seq: '부모 코드 순번'
                  use_yn: '사용여부 (Y ,N)'
                  ord: '정렬 순서'
                  remark: "메모"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
    delete:
      tags:
        - 코드 관리 컨트롤러
      summary: 코드 삭제
      operationId: codeDelete
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/CodeMngVO'
            examples:
              test:
                value:
                  code_seq: '13'
                  flag : 'd'
              guide:
                value:
                  code: '코드 순번'
                  flag: '해당 코드 : d , 코드와 자식코드 전체 : dp '
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
  /api/mng/code/list:
    get:
      tags:
        - 코드 관리 컨트롤러
      summary: 코드 리스트
      operationId: codeList
      parameters:
        - name: codeVO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CodeMngVO'
          examples:
            test:
              value:
                code: "test"
            guide:
              value:
                code: "코드"
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
  /api/mng/code/detail/{code}:
    get:
      tags:
        - 코드 관리 컨트롤러
      summary: 코드 상세정보
      operationId: codeDetailAjax
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
  /api/mng/code/check:
    post:
      tags:
        - 코드 관리 컨트롤러
      summary: 코드 중복체크
      operationId: codeCheck
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/CodeMngVO'
            examples:
              test:
                value:
                  code: "test"
              guide:
                value:
                  code: "코드"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
components:
  schemas:
...
    ResultVO:
      type: object
      properties:
        resultMsg:
          type: string
          description: 응답 메시지
        resultCode:
          type: integer
          description: 응답 코드
          format: int32
        result:
          type: object
          additionalProperties:
            type: object
            description: 응답 결과 데이터
          description: 응답 결과 데이터
      description: REST API 응답 결과 객체
...
    CodeMngVO:
      type: object
      properties:
        searchCondition:
          type: string
          description: 검색 조건
        searchKeyword:
          type: string
          description: 검색어
        searchUseYn:
          type: string
          description: 검색 사용 여부
          example: Y or N
        searchOrderBy:
          type: string
          description: 검색 정렬 기준
        pageIndex:
          type: integer
          description: 현재 페이지
          format: int32
        pageUnit:
          type: integer
          description: 페이지 개수
          format: int32
        file:
          type: array
          description: 파일 배열
          items:
            type: string
            format: binary
        code_seq:
          type: string
        code:
          type: string
        code_nm:
          type: string
        prt_seq:
          type: string
        use_yn:
          type: string
        remark:
          type: string
        inpt_seq:
          type: string
        upd_seq:
          type: string
        inpt_date:
          type: string
        ord:
          type: string
        flag:
          type: string
        code_check_yn:
          type: string
        id:
          type: string
        parent:
          type: string
        text:
          type: string
        state:
          type: object
          additionalProperties:
            type: boolean
        code_detail:
          type: string
        code_detail_nm:
          type: string
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT